+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- Clean-up handling of images, it was really messy.
 
 ## [2.4.1] - 2022-08-20
 ### Fixed